+++ notifier.go
@@ -17,7 +17,7 @@ import (
 )
 
 const notifierName = "gobrake"
-const notifierVersion = "4.2.0"
+const notifierVersion = "5.0.0"
 const userAgent = notifierName + "/" + notifierVersion
 
 const waitTimeout = 5 * time.Second