+++ CHANGELOG.md
@@ -12,6 +12,8 @@ Gobrake Changelog
   ([#148](https://github.com/airbrake/gobrake/pull/148))
 * Added the `APMHost` option that sets the host to which APM data should be sent
   to ([#150](https://github.com/airbrake/gobrake/pull/150))
+* Added support for remote configuration
+* Added support `go 1.15` ([#168](https://github.com/airbrake/gobrake/pull/168))
 
 ### [v4.2.0][v4.2.0] (July 24, 2020)